pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lightcurvedb"
version = "0.0.2"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "pydantic-settings",
    "numpy",
    "pydantic_numpy",
    "psycopg_binary",
    "testcontainers",
    "loguru",
    "tqdm",
    "socat @ git+https://github.com/simonsobs/socat",
    "astropy",
    "uuid7==0.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "coverage",
    "pytest-cov",
    "pre-commit",
]

[project.scripts]
lightcurvedb-ephemeral = "lightcurvedb.cli.ephemeral:main"
lightcurvedb-setup = "lightcurvedb.cli.setup:main"
lightcurvedb-socat = "lightcurvedb.cli.socat:main"

[tool.ruff.lint]
extend-select = ["I"]

[tool.coverage.run]
source = [
    "lightcurvedb",
]